# api/docs/swagger.yaml
definitions:
  entities.Project:
    properties:
      created_at:
        example: "2022-06-05T14:26:02.302718+03:00"
        type: string
      description:
        example: Mock API for an online store for selling shoes
        type: string
      id:
        example: 8f9c71b8-b84e-4417-8408-a62274f65a08
        type: string
      name:
        example: Mock Stripe API
        type: string
      subdomain:
        example: stripe-mock-api
        type: string
      updated_at:
        example: "2022-06-05T14:26:10.303278+03:00"
        type: string
      user_id:
        example: user_2oeyIzOf9xxxxxxxxxxxxxx
        type: string
    required:
      - created_at
      - description
      - id
      - name
      - subdomain
      - updated_at
      - user_id
    type: object
  entities.ProjectEndpoint:
    properties:
      created_at:
        example: "2022-06-05T14:26:02.302718+03:00"
        type: string
      delay_in_milliseconds:
        example: 1000
        type: integer
      description:
        example: Mock API for an online store for the /v1/products endpoint
        type: string
      id:
        example: 8f9c71b8-b84e-4417-8408-a62274f65a08
        type: string
      project_id:
        example: 8f9c71b8-b84e-4417-8408-a62274f65a08
        type: string
      request_count:
        example: 100
        type: integer
      request_method:
        example: GET
        type: string
      request_path:
        example: /v1/products
        type: string
      response_body:
        example: '{"message": "Hello World","status": 200}'
        type: string
      response_code:
        example: 200
        type: integer
      response_headers:
        example: '[{"Content-Type":"application/json"}]'
        type: string
      subdomain:
        example: stripe-mock-api
        type: string
      updated_at:
        example: "2022-06-05T14:26:10.303278+03:00"
        type: string
      user_id:
        example: user_2oeyIzOf9xxxxxxxxxxxxxx
        type: string
    required:
      - created_at
      - delay_in_milliseconds
      - description
      - id
      - project_id
      - request_count
      - request_method
      - request_path
      - response_body
      - response_code
      - response_headers
      - subdomain
      - updated_at
      - user_id
    type: object
  requests.ProjectCreateRequest:
    properties:
      description:
        type: string
      name:
        type: string
      subdomain:
        type: string
    required:
      - description
      - name
      - subdomain
    type: object
  requests.ProjectEndpointStoreRequest:
    properties:
      delay_in_milliseconds:
        type: integer
      description:
        type: string
      request_method:
        type: string
      request_path:
        type: string
      response_body:
        type: string
      response_code:
        type: integer
      response_headers:
        type: string
    required:
      - delay_in_milliseconds
      - description
      - request_method
      - request_path
      - response_body
      - response_code
      - response_headers
    type: object
  requests.ProjectEndpointUpdateRequest:
    properties:
      delay_in_milliseconds:
        type: integer
      description:
        type: string
      request_method:
        type: string
      request_path:
        type: string
      response_body:
        type: string
      response_code:
        type: integer
      response_headers:
        type: string
    required:
      - delay_in_milliseconds
      - description
      - request_method
      - request_path
      - response_body
      - response_code
      - response_headers
    type: object
  requests.ProjectUpdateRequest:
    properties:
      description:
        type: string
      name:
        type: string
      subdomain:
        type: string
    required:
      - description
      - name
      - subdomain
    type: object
  responses.BadRequest:
    properties:
      data:
        example: The request body is not a valid JSON string
        type: string
      message:
        example: The request isn't properly formed
        type: string
      status:
        example: error
        type: string
    required:
      - data
      - message
      - status
    type: object
  responses.InternalServerError:
    properties:
      message:
        example: We ran into an internal error while handling the request.
        type: string
      status:
        example: error
        type: string
    required:
      - message
      - status
    type: object
  responses.NoContent:
    properties:
      message:
        example: item deleted successfully
        type: string
      status:
        example: success
        type: string
    required:
      - message
      - status
    type: object
  responses.NotFound:
    properties:
      message:
        example: cannot find item with ID [32343a19-da5e-4b1b-a767-3298a73703ca]
        type: string
      status:
        example: error
        type: string
    required:
      - message
      - status
    type: object
  responses.Ok-array_entities_Project:
    properties:
      data:
        items:
          $ref: "#/definitions/entities.Project"
        type: array
      message:
        example: Request handled successfully
        type: string
      status:
        example: success
        type: string
    required:
      - data
      - message
      - status
    type: object
  responses.Ok-array_entities_ProjectEndpoint:
    properties:
      data:
        items:
          $ref: "#/definitions/entities.ProjectEndpoint"
        type: array
      message:
        example: Request handled successfully
        type: string
      status:
        example: success
        type: string
    required:
      - data
      - message
      - status
    type: object
  responses.Ok-entities_Project:
    properties:
      data:
        $ref: "#/definitions/entities.Project"
      message:
        example: Request handled successfully
        type: string
      status:
        example: success
        type: string
    required:
      - data
      - message
      - status
    type: object
  responses.Ok-entities_ProjectEndpoint:
    properties:
      data:
        $ref: "#/definitions/entities.ProjectEndpoint"
      message:
        example: Request handled successfully
        type: string
      status:
        example: success
        type: string
    required:
      - data
      - message
      - status
    type: object
  responses.Unauthorized:
    properties:
      data:
        example:
          Make sure your Bearer token is set in the [Bearer] header in the
          request
        type: string
      message:
        example: You are not authorized to carry out this request.
        type: string
      status:
        example: error
        type: string
    required:
      - data
      - message
      - status
    type: object
  responses.UnprocessableEntity:
    properties:
      data:
        additionalProperties:
          items:
            type: string
          type: array
        type: object
      message:
        example: validation errors while sending message
        type: string
      status:
        example: error
        type: string
    required:
      - data
      - message
      - status
    type: object
host: api.httpmock.dev
info:
  contact:
    email: arnold@httpmock.dev
    name: Acho Arnold
  description: Reference for the httpmock.dev API.
  license:
    name: AGPL-3.0
    url: https://raw.githubusercontent.com/NdoleStudio/httpmock/main/LICENSE
  title: HTTP Mock API Reference
  version: "1.0"
paths:
  /v1/projects:
    get:
      description:
        Fetches the list of all projects available to the currently authenticated
        user
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/responses.Ok-array_entities_Project"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/responses.BadRequest"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/responses.Unauthorized"
        "422":
          description: Unprocessable Entity
          schema:
            $ref: "#/definitions/responses.UnprocessableEntity"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/responses.InternalServerError"
      security:
        - BearerAuth: []
      summary: List of projects
      tags:
        - Projects
    post:
      description: This endpoint creates a new project for a user
      parameters:
        - description: project create payload
          in: body
          name: payload
          required: true
          schema:
            $ref: "#/definitions/requests.ProjectCreateRequest"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/responses.Ok-entities_Project"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/responses.BadRequest"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/responses.Unauthorized"
        "422":
          description: Unprocessable Entity
          schema:
            $ref: "#/definitions/responses.UnprocessableEntity"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/responses.InternalServerError"
      security:
        - BearerAuth: []
      summary: Store a project
      tags:
        - Projects
  /v1/projects/:projectId/endpoints:
    get:
      description:
        Fetches the list of all projects endpoints available to the currently
        authenticated user and project
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/responses.Ok-array_entities_ProjectEndpoint"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/responses.BadRequest"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/responses.Unauthorized"
        "422":
          description: Unprocessable Entity
          schema:
            $ref: "#/definitions/responses.UnprocessableEntity"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/responses.InternalServerError"
      security:
        - BearerAuth: []
      summary: List of project endpoints
      tags:
        - ProjectEndpoints
    post:
      description: This endpoint stores a new project endpoint for a user
      parameters:
        - description: project endpoint store payload
          in: body
          name: payload
          required: true
          schema:
            $ref: "#/definitions/requests.ProjectEndpointStoreRequest"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/responses.Ok-entities_ProjectEndpoint"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/responses.BadRequest"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/responses.Unauthorized"
        "422":
          description: Unprocessable Entity
          schema:
            $ref: "#/definitions/responses.UnprocessableEntity"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/responses.InternalServerError"
      security:
        - BearerAuth: []
      summary: store a new project endpoint
      tags:
        - ProjectEndpoints
  /v1/projects/{projectId}:
    delete:
      description: This endpoint deletes a project
      parameters:
        - description: Project ID
          in: path
          name: projectId
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/responses.NoContent"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/responses.BadRequest"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/responses.Unauthorized"
        "404":
          description: Not Found
          schema:
            $ref: "#/definitions/responses.NotFound"
        "422":
          description: Unprocessable Entity
          schema:
            $ref: "#/definitions/responses.UnprocessableEntity"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/responses.InternalServerError"
      security:
        - BearerAuth: []
      summary: Delete a project
      tags:
        - Projects
    get:
      description: This endpoint gets a project for a user
      parameters:
        - description: Project ID
          in: path
          name: projectId
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/responses.Ok-entities_Project"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/responses.BadRequest"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/responses.Unauthorized"
        "422":
          description: Unprocessable Entity
          schema:
            $ref: "#/definitions/responses.UnprocessableEntity"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/responses.InternalServerError"
      security:
        - BearerAuth: []
      summary: Get a project
      tags:
        - Projects
    put:
      description: This endpoint updates a project for a user
      parameters:
        - description: Project ID
          in: path
          name: projectId
          required: true
          type: string
        - description: project update payload
          in: body
          name: payload
          required: true
          schema:
            $ref: "#/definitions/requests.ProjectUpdateRequest"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/responses.Ok-entities_Project"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/responses.BadRequest"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/responses.Unauthorized"
        "422":
          description: Unprocessable Entity
          schema:
            $ref: "#/definitions/responses.UnprocessableEntity"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/responses.InternalServerError"
      security:
        - BearerAuth: []
      summary: Update a project
      tags:
        - Projects
  /v1/projects/{projectId}/endpoints/{projectEndpointId}:
    delete:
      description: This API deletes a project endpoint for a user
      parameters:
        - description: Project ID
          in: path
          name: projectId
          required: true
          type: string
        - description: Project Endpoint ID
          in: path
          name: projectEndpointId
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/responses.NoContent"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/responses.BadRequest"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/responses.Unauthorized"
        "404":
          description: Not Found
          schema:
            $ref: "#/definitions/responses.NotFound"
        "422":
          description: Unprocessable Entity
          schema:
            $ref: "#/definitions/responses.UnprocessableEntity"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/responses.InternalServerError"
      security:
        - BearerAuth: []
      summary: Delete a project endpoint
      tags:
        - ProjectEndpoints
    get:
      description: This URL gets a project endpoint for a user
      parameters:
        - description: Project ID
          in: path
          name: projectId
          required: true
          type: string
        - description: Project Endpoint ID
          in: path
          name: projectEndpointId
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/responses.Ok-entities_ProjectEndpoint"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/responses.BadRequest"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/responses.Unauthorized"
        "422":
          description: Unprocessable Entity
          schema:
            $ref: "#/definitions/responses.UnprocessableEntity"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/responses.InternalServerError"
      security:
        - BearerAuth: []
      summary: Get a project endpoint
      tags:
        - ProjectEndpoints
    put:
      description: This endpoint updates a project endpoint for a user
      parameters:
        - description: Project ID
          in: path
          name: projectId
          required: true
          type: string
        - description: Project Endpoint ID
          in: path
          name: projectEndpointId
          required: true
          type: string
        - description: project update payload
          in: body
          name: payload
          required: true
          schema:
            $ref: "#/definitions/requests.ProjectEndpointUpdateRequest"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/responses.Ok-entities_ProjectEndpoint"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/responses.BadRequest"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/responses.Unauthorized"
        "422":
          description: Unprocessable Entity
          schema:
            $ref: "#/definitions/responses.UnprocessableEntity"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/responses.InternalServerError"
      security:
        - BearerAuth: []
      summary: Update a project endpoint
      tags:
        - ProjectEndpoints
schemes:
  - https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"